% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot_pampa.f}
\alias{barplot_pampa.f}
\title{Barplot}
\usage{
barplot_pampa.f(
  agregation,
  metrique,
  factGraph = NULL,
  factGraphSel = NA,
  listFact,
  listFactSel = NA,
  tableMetrique,
  new_window = TRUE,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{agregation}{chr, type d'agregation, par espece (espece) ou par groupe d'espece (unitobs)}

\item{metrique}{chr, metrique choisie}

\item{factGraph}{chr, facteur de separation des graphique}

\item{factGraphSel}{chr, selection de modalites du facteur de separation des graphiques}

\item{listFact}{chr, facteur(s) de regroupement}

\item{listFactSel}{list, modalite selectionnees pour le(s) facteur(s) de regroupement}

\item{tableMetrique}{chr, nom de la table de metrique}

\item{new_window}{bool, affichage du graphique dans une nouvelle fenetre ?}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}
}
\description{
Creation d'un barplot selon les parametres fournis
}
\details{
Fonction pour tracer un barplot.

\code{agregation} peut prendre la valeur 'espece' ou 'unitobs'. Si 'espece', l'agregation se fera
par especes, si unitobs, elle se fera par groupe d'especes.

\code{tableMetrique} peut prendre les valeurs 'unitSpSz' (/station /especes /classe de taille),
'unitSp' (/station /especes) ou 'unit' (de biodiversite (/station)). Ce dernier cas n'est
possible uniquement lorsque \code{agregation == 'unitobs'}. La table de metriques 'unitSpSz'
n'est pas disponible si le jeu de donnees est un jeu de donnees de benthos.

\code{metrique} peut prendre differentes valeurs en fonction de celle de \code{tableMetrique}
 et du jeu de donnees.

\code{factGraphSel} peut prendre differentes valeurs en fonction de celle de \code{factGraph}. Ce
parametre est facultatif. S'il est egal a NA, toutes les modalites du facteur seront
selectionnees.

\code{listFact} ne peut pas avoir plus de 2 facteurs.

\code{listFactSel} peut prendre differentes valeurs en fonction de celle de \code{listFact}.
Ce parametre est facultatif. S'il est egal a NA, alors toutes les modalites du ou des facteur(s)
de regroupement selectionne(s) seront prises en compte.
}
\examples{
barplot_pampa.f(
  agregation = "espece",
  metrique = "density",
  factGraph = "family",
  factGraphSel = "Acanthuridae",
  listFact = c("year", "protection.status"),
  listFactSel = NA,
  tableMetrique = "unitSp",
  new_window = TRUE,
  dataEnv = .dataEnv, baseEnv = .baseEnv)

}
