% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_occurence.R
\name{freq_occurrence.f}
\alias{freq_occurrence.f}
\title{Frequences d'occurrence}
\usage{
freq_occurrence.f(
  agregation,
  factGraph = NULL,
  factGraphSel = NA,
  listFact,
  listFactSel = NA,
  new_window = TRUE,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{agregation}{chr, type d'agregation, par espece (espece) ou par groupe d'espece (unitobs)}

\item{factGraph}{chr, facteur de separation des graphique.}

\item{factGraphSel}{chr, selection de modalites du facteur de separation des graphiques}

\item{listFact}{chr, facteur(s) de regroupement}

\item{listFactSel}{list, modalite selectionnees pour le(s) facteur(s) de regroupement}

\item{new_window}{bool, affichage du graphique dans une nouvelle fenetre ?}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}
}
\description{
Creation d'un barplot de frequence d'occurrence selon les parametres fournis
}
\details{
Fonction pour tracer un barplot de frequences d'occurrence.

\code{agregation} peut prendre la valeur 'espece' ou 'unitobs'. Si 'espece', l'agregation se fera
par especes, si unitobs, elle se fera par groupe d'especes.

\code{factGraphSel} peut prendre differentes valeurs en fonction de celle de \code{factGraph}. Ce
parametre est facultatif. S'il est egal a NA, toutes les modalites du facteur seront
selectionnees.

\code{listFact} ne peut pas avoir plus de 2 facteurs.

\code{listFactSel} peut prendre differentes valeurs en fonction de celle de \code{listFact}. Ce
parametre est facultatif. S'il est egal a NA, alors toutes les modalites du ou des facteur(s) de
regroupement selectionne(s) seront prises en compte.
}
\examples{
freq_occurrence.f(
  agregation = "espece",
  factGraph = "species.code",
  factGraphSel = "Acanoliv",
  listFact = c("year", "protection.status"),
  listFactSel = NA,
  new_window = TRUE,
  dataEnv = .dataEnv, baseEnv = .baseEnv)

}
