% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_familles.R
\name{freq_occurrence_familles.f}
\alias{freq_occurrence_familles.f}
\title{Barplot des frequences de familles}
\usage{
freq_occurrence_familles.f(
  factGraph,
  factGraphSel = NA,
  fact,
  factSel = NA,
  families = NA,
  new_window = TRUE,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{factGraphSel}{chr, selection de modalites du premier facteur explicatif}

\item{fact}{chr, second facteur explicatif (facultatif)}

\item{factSel}{chr, modalites selectionnees pour le second facteur explicatif}

\item{families}{chr, familles a prendre en compte}

\item{new_window}{bool, affichage du graphique dans une nouvelle fenetre ?}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}

\item{factGaph}{chr, premier facteur explicatif}
}
\description{
Creation d'un barplot de frequences d'occurence pour les familles
}
\details{
Fonction pour tracer un barplot de frequences d'occurrence de familles.

\code{factGraphSel} peut prendre differentes valeurs en fonction de celle de \code{factGraph}.
Ce parametre est facultatif. S'il est egal a NA, toutes les modalites du facteur seront
selectionnees.

\code{factSel} peut prendre differentes valeurs en fonction de celle de \code{fact}. Ce
parametre est facultatif. S'il est egal a NA, alors toutes les modalites du facteur de
regroupement selectionne seront prises en compte.

\code{families} correspond aux familles que l'on veut etudier. Si ce parametre est egal a NA,
alors toutes les familles seront prises en compte
}
\examples{
PAMPA::freq_occurrence_families.f(
  factGraph = "protection.status",
  factGraphSel = NA,
  fact = "year",
  factSel = NA,
  families = NA,
  new_window = TRUE,
  dataEnv = .dataEnv, baseEnv = .baseEnv)

}
