% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_files.f}
\alias{load_files.f}
\title{Importation des donnees et calculs des poids}
\usage{
load_files.f(filePathes, dminMax = 5, dataEnv, baseEnv)
}
\arguments{
\item{filePathes}{chr, vecteur nomme contenant les chemins des fichiers :
\itemize{unitobs}{ : chr, fichier unitobs}
\itemize{obs}{ : chr, fichier d'observation}
\itemize{refesp}{ : chr, referentiel especes}
\itemize{locrefesp}{ : chr, referentiel especes local (facultatif)}
\itemize{refspa}{ : chr, referentiel spatial (facultatif)}
\itemize{ws}{ : chr, dossier de travail}
\itemize{results}{ : chr, dossier ou enregistrer les donnees (facultatif)}
\itemize{...}}

\item{dminMax}{seuil (en m) au-dessus duquel les observations ne sont pas prises en compte}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}
}
\value{
list, liste nommee contenant les tables de donnees :
\itemize{obs}{ : donnees d'observation}
\itemize{unitobs}{ : unites d'observation}
\itemize{refesp}{ : referentiel espece}
}
\description{
Chargement des donnees depuis les chemins des fichiers et calcul des poids.
}
\details{
Cette fonction permet de faire le chargement des donnees.

Elle prend en entree les chemins des differents fichiers (sous forme d'un
vecteur nomme).

Le referentiel espece local et le referentiel statial sont facultatifs.

Le chemin du dossier "results" est facultatif egalement. S'il n'est pas renseigne, les donnees
seront enregistrees dans un dossier nomme 'results' dans le dossier de travail.
Elle prend egalement en entree deux environnements, celui où stocker les donnees, et
l'environnement parent. (Pas forcement tres pratique d'utilisation, a enlever plus tard ?)

Le dossier de travail doit etre un dossier nomme 'Data' et doit etre le dossier de stockage
des donnees.

Cette fonction charge les donnees a partir des chemins de fichiers donnes. Elle calcule ensuite
les poids pour chaque AMP. Puis, elle calcule les tables de metriques, qu'elle enregistre dans
l'environnement dataEnv afin de pouvoir les réutiliser plus tard. Elle renvoie ensuite les
differentes tables de donnees, sous forme d'une liste nommee.
}
\examples{
ws_path <- system.file('example_data/COTE BLEUE/Data/', package = 'PAMPA')
unitobs_path <- system.file('example_data/COTE BLEUE/Data/UnitObs_Staviro_CB191110_280720.txt', package = 'PAMPA')
obs_path <- system.file('example_data/COTE BLEUE/Data/Obs_Staviro_CB191110_AllIdentifiedSpecies_280720.txt', package = 'PAMPA')
refesp_path <- system.file('example_data/COTE BLEUE/Data/refEspecesMed_general_270720 - Int peche est intCH_CB.txt', package = 'PAMPA')

filepaths <- c(unitobs = unitobs_path, obs = obs_path, refesp =  refesp_path, refspa = NULL, ws = ws_path)

load_files.f(filepaths, dminMax = 5, .dataEnv, .baseEnv)

}
