% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_descriptives.R
\name{occurence_especes_site_annee.f}
\alias{occurence_especes_site_annee.f}
\title{Occurence des especes par site et par annee}
\usage{
occurence_especes_site_annee.f(
  site = NA,
  year = NA,
  species = NA,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{site}{chr, selection des sites a etudier, si NA, tous les sites seront selectionnes}

\item{year}{chr, selection des annees a etudier, si NA toutes les annees seront selectionnees}

\item{species}{chr, selection des especes a etudier, si NA toutes les especes seront selectionnees}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}
}
\description{
Cree un graph (et une carte si refspa disponible) presentant le nombre d'occurrence
des especes par site et par annee
}
\examples{
PAMPA::occurence_especes_site_annee.f(site = NA, year = NA,
  species = c("Acanthurus_sp2", "Chaetodon_kleinii", "Naso_tonganus", "Scaridae"),
  .dataEnv, .baseEnv)

}
