% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_data.R
\name{restoreData.f}
\alias{restoreData.f}
\title{Restauration des donnees}
\usage{
restoreData.f(dataEnv, baseEnv)
}
\arguments{
\item{dataEnv}{: environnement de stockage des donnees}

\item{baseEnv}{: environnement parent}
}
\value{
list, liste contenant les tables de donnees selon la selection.
 \itemize{unitobs}{ : chr, fichier unitobs}
 \itemize{obs}{ : chr, fichier d'observation}
 \itemize{refesp}{ : chr, referentiel especes}
}
\description{
Permet la restauration des donnees.
}
\details{
Fonction pour la restauration des donnees apres la selection et le recalcul.

Permet de faire la restauration des donnees. Cela revient a re-charger les donnees, mais en plus
rapide.
}
\examples{
restoreData.f(.dataEnv, .baseEnv)

}
