% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{selection.f}
\alias{selection.f}
\title{Selection des donnees}
\usage{
selection.f(field, dataEnv, baseEnv)
}
\arguments{
\item{field}{: chr, unitobs ou refesp, type de champs choisi pour la selection.}

\item{dataEnv}{: environnement de stockage des donnees}

\item{baseEnv}{: environnement parent}
}
\value{
list, liste contenant les tables de donnees selon la selection.
 \itemize{unitobs}{ : chr, fichier unitobs}
 \itemize{obs}{ : chr, fichier d'observation}
 \itemize{refesp}{ : chr, referentiel especes}
}
\description{
Selection des donnees selon un champs du referentiel espece ou un
champs des unitobs.
}
\details{
Cette fonction permet de faire la selection et le recalcul des donnees.

Elle prend en entree le type de critere selon lequel le recalcul sera fait
(champs du referentiel espece ou champs des unitobs).

Elle prend egalement en entree deux environnements, celui de stockage des donnees
et l'environnement parent.

Elle va ensuite faire la selection des donnees, en demandant a l'utilisateur
selon quel(s) critere(s) du champs choisi il veut la faire.

Elle retourne la table des observations modifiee, celle des uniotobs, et celle du referentiel
spatial.
}
\examples{
ws_path <- 'inst/example_data/'
unitobs_path <- 'inst/example_data/Data/UnitObs_Staviro_Example.txt'
obs_path <- 'inst/example_data/Data/Obs_Staviro_Example.txt'
refesp_path <- 'inst/example_data/Data/SpeciesReferenceTable_Example.txt'

filePathes <- c(unitobs = unitobs_path, obs = obs_path, refesp =  refesp_path, ws = ws_path)

data <- PAMPA::load_files.f(filePathes, dminMax = 5, .dataEnv, .baseEnv)

data_sel <- PAMPA::selection.f('unitobs', .dataEnv, .baseEnv)

}
