% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{WP2barplot.esp.f}
\alias{WP2barplot.esp.f}
\title{Barplot (données agrégées par espèces)}
\usage{
WP2barplot.esp.f(
  metrique,
  factGraph,
  factGraphSel,
  listFact,
  listFactSel,
  tableMetrique,
  new_window = TRUE,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{metrique}{chr, métrique choisie}

\item{factGraph}{chr, facteur de séparation des graphiques}

\item{factGraphSel}{chr, sélection de modalité pour le facteur de séparation des graphiques}

\item{listFact}{chr,  facteur(s) de regroupement}

\item{listFactSel}{list, modalité sélectionnées pour le(s) facteur(s) de regroupement}

\item{tableMetrique}{chr, nom de la table de métrique}

\item{dataEnv}{environnement de stockage des données}

\item{baseEnv}{environnement parent}
}
\value{
none
}
\description{
Cette fonction est appelée par selectionVariables.f. Elle permet la création de
barplot pour des données agrégées par espèces selon les paramètres définis par l'utilisateur.
}
