% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrt.R
\name{WP2MRT.esp.f}
\alias{WP2MRT.esp.f}
\title{Arbres de régression multivariée (données agrégées par espèces)}
\usage{
WP2MRT.esp.f(
  metrique,
  factGraph,
  factGraphSel,
  listFact,
  listFactSel,
  tableMetrique,
  new_window = TRUE,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{metrique}{chr, métrique choisie}

\item{factGraph}{chr, facteur de séparation des graphiques}

\item{factGraphSel}{chr, sélection des modalité pour le facteur de séparation des graphiques}

\item{listFact}{chr, facteur(s) de regroupement}

\item{listFactSel}{list, modalités sélectionnées pour le(s) facteur(s) de regroupement}

\item{tableMetrique}{chr, nom de la table de métriques}

\item{dataEnv}{environnement de stockage des données}

\item{baseEnv}{environnement parent}
}
\value{
none
}
\description{
Cette fonction permet de faire des arbres de régression multivariée des données
agrégées par espèces
}
