% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_files.f}
\alias{load_files.f}
\title{Importation des donnees et calculs des poids}
\usage{
load_files.f(filePathes, dminMax = 5, dataEnv, baseEnv)
}
\arguments{
\item{filePathes}{chr, vecteur nomme contenant les chemins des fichiers :
\itemize{unitobs}{ : chr, fichier unitobs}
\itemize{obs}{ : chr, fichier d'observation}
\itemize{refesp}{ : chr, referentiel especes}
\itemize{locrefesp}{ : chr, referentiel especes local (facultatif)}
\itemize{refspa}{ : chr, referentiel spatial (facultatif)}
\itemize{ws}{ : chr, dossier de travail}
\itemize{results}{ : chr, dossier ou enregistrer les donnees (facultatif)}
\itemize{...}}

\item{dminMax}{seuil (en m) au-dessus duquel les observations ne sont pas prises en compte}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}
}
\value{
list, liste nommee contenant les tables de donnees :
\itemize{obs}{ : donnees d'observation}
\itemize{unitobs}{ : unites d'observation}
\itemize{refesp}{ : referentiel espece}
}
\description{
Chargement des donnees depuis les chemins des fichiers et calcul des poids.
}
\details{
Cette fonction permet de faire le chargement des donnees.

Elle prend en entree les chemins des differents fichiers (sous forme d'un
vecteur nomme).

Le referentiel espece local et le referentiel statial sont facultatifs.

Le chemin du dossier "Results" est facultatif egalement. S'il n'est pas renseigne, les donnees
seront enregistrees dans un dossier nomme 'Results' dans le dossier de travail.
Elle prend egalement en entree deux environnements, celui ou stocker les donnees, et
l'environnement parent. Ces environnements peuvent tous les deux correspondre a l'environnement
global (.GlobalEnv).

Le dossier de travail doit etre le dossier de stockage des donnees. Il s'agit d'un dossier
contenant un dossier nomme 'Data' ou sont stockees les donnees.

Cette fonction charge les donnees a partir des chemins de fichiers donnes. Elle calcule ensuite
les poids pour chaque AMP. Puis, elle calcule les tables de metriques, qu'elle enregistre dans
l'environnement dataEnv afin de pouvoir les reutiliser plus tard. Elle renvoie ensuite les
differentes tables de donnees, sous forme d'une liste nommee.
}
\examples{
# Creation d'environnements
.baseEnv <- environment()
.dataEnv <- new.env()

assign(".baseEnv", .baseEnv, .GlobalEnv)
assign(".dataEnv", .dataEnv, .GlobalEnv)

# Definition des chemins de l'environnement de travail et des fichiers
ws_path <- 'inst/example_data/'
unitobs_path <- 'inst/example_data/Data/UnitObs_Staviro_Example.txt'
obs_path <- 'inst/example_data/Data/Obs_Staviro_Example.txt'
refesp_path <- 'inst/example_data/Data/SpeciesReferenceTable_Example.txt'

filePathes <- c(unitobs = unitobs_path, obs = obs_path, refesp =  refesp_path, ws = ws_path)

# Lancement du chargement
data <- PAMPA::load_files.f(filePathes, dminMax = 5, .dataEnv, .baseEnv)

}
