% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{permanova_pampa.f}
\alias{permanova_pampa.f}
\title{Permanova}
\usage{
permanova_pampa.f(
  metric_table,
  metric,
  fact = NULL,
  formula = NULL,
  method = "bray",
  nb_perm = 1000,
  square_roots = FALSE,
  post_hoc = TRUE,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{metric_table}{chr, le nom de la table de metrique}

\item{metric}{chr, donnees etudiees ("pres_abs" (presence-absence), "abundance" (abondance) ou "density" (densite))}

\item{fact}{chr, liste des facteurs explicatifs. Utilise si aucune formule n'est renseignee.
La fomule est cree sous la forme \code{dissimilarity_matrix ~ fact1 * fact2 * ... * factn}.}

\item{formula}{chr, RHS de la formule du modele. Si elle est laissee nulle (la valeur par defaut),
alors les facteurs explicatifs sont utilises pour creer la formule.}

\item{method}{chr, methode utilisee pour la creation de la matrice de dissimilarites}

\item{nb_perm}{int, nombre de permutations}

\item{square_roots}{booleen, utilisation des square roots ?}

\item{post_hoc}{booleen, faire un post hoc apres la permanova ?}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}
}
\description{
Calcul d'une permanova selon les parametres fournis
}
\details{
Cette fonction permet de faire une permanova.
}
\examples{
PAMPA::permanova_pampa.f(
  metric_table = "unitSp",
  metric = "pres_abs",
  fact = c("site", "year"),
  method = "bray",
  nb_perm = 1000,
  square_roots = FALSE,
  post_hoc = FALSE,
  dataEnv = .dataEnv, baseEnv = .baseEnv)

}
